requirements.txt
llama-cpp-python==0.1.83
telebot~=0.0.5